Revert "feat: add WebRTC P2P mode for serverless gameplay"
