+++ src/components/ui/GameOverlay.tsx
@@ -5,7 +5,7 @@ import { StockPanel } from './StockPanel';
 import { PlayerList } from './PlayerList';
 import { playVictory, playDefeat, playClick, playPlaceUnit, playError } from '@/audio/sounds';
 import { canPlaceUnits } from '@/game/battle';
-import { unifiedGameClient } from '@/network/unifiedClient';
+import { gameClient } from '@/network/client';
 
 const overlayStyle: React.CSSProperties = {
   position: 'absolute',
@@ -94,7 +94,7 @@ export function GameOverlay() {
 
     if (spendMoney(gameState.unitCost)) {
       playPlaceUnit();
-      unifiedGameClient.send({ type: 'placeUnits', coords: lastClickedHex });
+      gameClient.send({ type: 'placeUnits', coords: lastClickedHex });
     } else {
       playError();
     }
