--- src/game/engine.ts
@@ -1,634 +1,0 @@
-/**
- * Shared Game Engine - Core game logic used by both server and P2P host
- * This module contains all game state management, player handling, and game loop logic
- */
-
-import type { GameState, Player, Country, Company, Faction, HexCoord } from '@/types/game';
-import type { ClientMessage, ServerMessage, GameInfo } from '@/network/messages';
-import { generateHexGrid } from '@/game/hex';
-import { processBattle, calculateUnitCost, canPlaceUnits, checkWinner } from '@/game/battle';
-import { createCompanies, updateStockPrice } from '@/game/stock';
-import { GAME } from '@/game/constants';
-
-export const AI_NAMES = [
-  'Admiral Zex',
-  'Commander Nova',
-  'Captain Orion',
-  'General Nebula',
-  'Marshal Cosmos',
-  'Warlord Pulsar',
-];
-
-export const AI_ACTION_INTERVAL = 800; // ms between AI actions
-
-export interface GameClient {
-  playerId: string;
-  playerName: string | null;
-  currentGameId: string | null;
-}
-
-export interface AiPlayer {
-  id: string;
-  name: string;
-  gameId: string;
-  lastActionTime: number;
-}
-
-export interface GameRoom {
-  id: string;
-  hostId: string;
-  gameState: GameState | null;
-  clients: Set<string>; // playerIds
-  aiPlayers: Map<string, AiPlayer>; // aiId -> AiPlayer
-  factionCounter: number;
-  gameLoop: ReturnType<typeof setInterval> | null;
-}
-
-/**
- * Transport interface - implemented differently by WebSocket server and P2P host
- */
-export interface GameTransport {
-  send(playerId: string, msg: ServerMessage): void;
-  broadcast(msg: ServerMessage): void;
-  broadcastToGame(room: GameRoom, msg: ServerMessage): void;
-  getClient(playerId: string): GameClient | undefined;
-}
-
-/**
- * Core game engine that handles all game logic
- */
-export class GameEngine {
-  private transport: GameTransport;
-  private clients: Map<string, GameClient>;
-  private games: Map<string, GameRoom>;
-  private playerCounter: number;
-  private gameCounter: number;
-  private aiCounter: number;
-  private onGameListChanged?: () => void;
-
-  constructor(
-    transport: GameTransport,
-    clients: Map<string, GameClient>,
-    games: Map<string, GameRoom>,
-    counters: { playerCounter: number; gameCounter: number; aiCounter: number }
-  ) {
-    this.transport = transport;
-    this.clients = clients;
-    this.games = games;
-    this.playerCounter = counters.playerCounter;
-    this.gameCounter = counters.gameCounter;
-    this.aiCounter = counters.aiCounter;
-  }
-
-  setOnGameListChanged(callback: () => void): void {
-    this.onGameListChanged = callback;
-  }
-
-  getNextPlayerId(): string {
-    return `player${++this.playerCounter}`;
-  }
-
-  getPlayerCounter(): number {
-    return this.playerCounter;
-  }
-
-  getGameCounter(): number {
-    return this.gameCounter;
-  }
-
-  getAiCounter(): number {
-    return this.aiCounter;
-  }
-
-  handleMessage(playerId: string, msg: ClientMessage): void {
-    const client = this.clients.get(playerId);
-    if (!client) return;
-
-    switch (msg.type) {
-      case 'setName':
-        client.playerName = msg.playerName;
-        break;
-      case 'listGames':
-        this.handleListGames(client);
-        break;
-      case 'createGame':
-        this.handleCreateGame(client);
-        break;
-      case 'joinGame':
-        this.handleJoinGame(client, msg.gameId);
-        break;
-      case 'leaveGame':
-        this.handleLeaveGame(client);
-        break;
-      case 'startGame':
-        this.handleStartGame(client);
-        break;
-      case 'placeUnits':
-        this.handlePlaceUnits(client, msg.coords);
-        break;
-      case 'ping':
-        this.transport.send(client.playerId, { type: 'pong', timestamp: msg.timestamp });
-        break;
-      case 'addAi':
-        this.handleAddAi(client);
-        break;
-    }
-  }
-
-  handleDisconnect(playerId: string): void {
-    const client = this.clients.get(playerId);
-    if (client?.currentGameId) {
-      this.handleLeaveGame(client);
-    }
-    this.clients.delete(playerId);
-  }
-
-  private handleListGames(client: GameClient): void {
-    const games: GameInfo[] = [];
-    for (const [id, room] of this.games) {
-      const host = this.clients.get(room.hostId);
-      const players = this.getGamePlayers(room);
-      games.push({
-        id,
-        hostName: host?.playerName ?? 'Unknown',
-        playerCount: players.length,
-        maxPlayers: 6,
-        phase: room.gameState?.phase ?? 'lobby',
-        players: players.map((p) => p.name),
-      });
-    }
-    this.transport.send(client.playerId, { type: 'gameList', games });
-  }
-
-  private handleCreateGame(client: GameClient): void {
-    if (!client.playerName) {
-      this.transport.send(client.playerId, { type: 'error', message: 'Set your name first' });
-      return;
-    }
-
-    if (client.currentGameId) {
-      this.handleLeaveGame(client);
-    }
-
-    const gameId = `game${++this.gameCounter}`;
-    const room: GameRoom = {
-      id: gameId,
-      hostId: client.playerId,
-      gameState: null,
-      clients: new Set([client.playerId]),
-      aiPlayers: new Map(),
-      factionCounter: 0,
-      gameLoop: null,
-    };
-
-    this.games.set(gameId, room);
-    client.currentGameId = gameId;
-
-    const factionId = `faction${room.factionCounter++}`;
-    this.transport.send(client.playerId, { type: 'joinedGame', gameId, factionId });
-    this.broadcastLobbyUpdate(room);
-    this.broadcastGameList();
-
-    console.log(`Game ${gameId} created by ${client.playerName}`);
-  }
-
-  private handleJoinGame(client: GameClient, gameId: string): void {
-    if (!client.playerName) {
-      this.transport.send(client.playerId, { type: 'error', message: 'Set your name first' });
-      return;
-    }
-
-    const room = this.games.get(gameId);
-    if (!room) {
-      this.transport.send(client.playerId, { type: 'error', message: 'Game not found' });
-      return;
-    }
-
-    if (room.gameState?.phase === 'playing') {
-      this.transport.send(client.playerId, { type: 'error', message: 'Game already in progress' });
-      return;
-    }
-
-    if (client.currentGameId) {
-      this.handleLeaveGame(client);
-    }
-
-    room.clients.add(client.playerId);
-    client.currentGameId = gameId;
-
-    const factionId = `faction${room.factionCounter++}`;
-    this.transport.send(client.playerId, { type: 'joinedGame', gameId, factionId });
-    this.broadcastLobbyUpdate(room);
-    this.broadcastGameList();
-
-    console.log(`${client.playerName} joined game ${gameId}`);
-  }
-
-  private handleLeaveGame(client: GameClient): void {
-    if (!client.currentGameId) return;
-
-    const room = this.games.get(client.currentGameId);
-    if (!room) {
-      client.currentGameId = null;
-      return;
-    }
-
-    room.clients.delete(client.playerId);
-    this.transport.send(client.playerId, { type: 'leftGame' });
-
-    // If game is in progress, remove player from game and neutralize their countries
-    if (room.gameState && room.gameState.phase === 'playing') {
-      const playerEntry = room.gameState.players.find((p) => p.id === client.playerId);
-      if (playerEntry) {
-        const factionId = playerEntry.factionId;
-
-        for (let i = 0; i < room.gameState.countries.length; i++) {
-          const country = room.gameState.countries[i];
-          const units = country.units;
-          const remainingUnits = units[factionId] || 0;
-          if (remainingUnits > 0) {
-            // Remove any lingering units of the departed faction and add them to neutral
-            delete units[factionId];
-            units.neutral = (units.neutral || 0) + remainingUnits;
-          }
-        }
-
-        // Recalculate unit cost and broadcast updated state
-        room.gameState.unitCost = calculateUnitCost(room.gameState.countries);
-      }
-    }
-
-    // If room is empty or host left, clean up
-    if (room.clients.size === 0 || room.hostId === client.playerId) {
-      this.destroyGame(room);
-    } else {
-      this.broadcastLobbyUpdate(room);
-    }
-
-    client.currentGameId = null;
-    this.broadcastGameList();
-
-    console.log(`${client.playerName} left game ${room.id}`);
-  }
-
-  private destroyGame(room: GameRoom): void {
-    if (room.gameLoop) {
-      clearInterval(room.gameLoop);
-    }
-
-    // Notify remaining clients
-    for (const playerId of room.clients) {
-      const c = this.clients.get(playerId);
-      if (c) {
-        c.currentGameId = null;
-        this.transport.send(c.playerId, { type: 'leftGame' });
-      }
-    }
-
-    this.games.delete(room.id);
-    console.log(`Game ${room.id} destroyed`);
-  }
-
-  private handleStartGame(client: GameClient): void {
-    if (!client.currentGameId) return;
-
-    const room = this.games.get(client.currentGameId);
-    if (!room) return;
-
-    // Only host can start
-    if (room.hostId !== client.playerId) {
-      this.transport.send(client.playerId, { type: 'error', message: 'Only host can start the game' });
-      return;
-    }
-
-    if (room.gameState?.phase === 'playing') return;
-
-    const players = this.getGamePlayers(room);
-    if (players.length === 0) return;
-
-    const factions: Faction[] = [
-      { id: 'neutral', name: 'Neutral' },
-      ...players.map((p) => ({ id: p.factionId, name: p.name })),
-    ];
-
-    const startingPositions: HexCoord[] = [
-      { q: GAME.MAP_RADIUS, r: 0 },
-      { q: -GAME.MAP_RADIUS, r: 0 },
-      { q: 0, r: GAME.MAP_RADIUS },
-      { q: 0, r: -GAME.MAP_RADIUS },
-      { q: GAME.MAP_RADIUS, r: -GAME.MAP_RADIUS },
-      { q: -GAME.MAP_RADIUS, r: GAME.MAP_RADIUS },
-    ];
-
-    const hexCoords = generateHexGrid(GAME.MAP_RADIUS);
-    const countries: Country[] = hexCoords.map((coords) => {
-      const playerIndex = startingPositions.findIndex(
-        (pos) => pos.q === coords.q && pos.r === coords.r
-      );
-
-      if (playerIndex !== -1 && playerIndex < players.length) {
-        return {
-          coords,
-          units: { [players[playerIndex].factionId]: 10 },
-          nextBattleTime: Date.now() + Math.random() * GAME.BATTLE_MAX_INTERVAL,
-        };
-      }
-
-      return {
-        coords,
-        units: {
-          neutral:
-            GAME.NEUTRAL_UNITS_MIN +
-            Math.floor(Math.random() * (GAME.NEUTRAL_UNITS_MAX - GAME.NEUTRAL_UNITS_MIN)),
-        },
-        nextBattleTime: Date.now() + Math.random() * GAME.BATTLE_MAX_INTERVAL,
-      };
-    });
-
-    const companies = createCompanies(GAME.STOCK_COUNT);
-
-    room.gameState = {
-      phase: 'playing',
-      countries,
-      companies,
-      factions,
-      players,
-      unitCost: calculateUnitCost(countries),
-      winner: null,
-    };
-
-    this.transport.broadcastToGame(room, { type: 'gameStarted' });
-    this.startGameLoop(room);
-    this.broadcastGameList();
-
-    console.log(`Game ${room.id} started with ${players.length} players`);
-  }
-
-  private handlePlaceUnits(client: GameClient, coords: HexCoord): void {
-    if (!client.currentGameId) return;
-
-    const room = this.games.get(client.currentGameId);
-    if (!room?.gameState || room.gameState.phase !== 'playing') return;
-
-    const player = this.getGamePlayers(room).find((p) => p.id === client.playerId);
-    if (!player) return;
-
-    const factionId = player.factionId;
-    const countryIndex = room.gameState.countries.findIndex(
-      (c: Country) => c.coords.q === coords.q && c.coords.r === coords.r
-    );
-
-    if (countryIndex === -1) return;
-
-    const country = room.gameState.countries[countryIndex];
-    if (!canPlaceUnits(country, room.gameState.countries, factionId)) return;
-
-    const newUnits = { ...country.units };
-    newUnits[factionId] = (newUnits[factionId] || 0) + 1;
-
-    // Reset battle timer when placing units (grace period)
-    room.gameState.countries[countryIndex] = {
-      ...country,
-      units: newUnits,
-      nextBattleTime: Date.now() + GAME.PLACEMENT_GRACE_PERIOD,
-    };
-
-    room.gameState.unitCost = calculateUnitCost(room.gameState.countries);
-  }
-
-  private handleAddAi(client: GameClient): void {
-    if (!client.currentGameId) return;
-
-    const room = this.games.get(client.currentGameId);
-    if (!room) return;
-
-    // Only host can add AI
-    if (room.hostId !== client.playerId) {
-      this.transport.send(client.playerId, { type: 'error', message: 'Only host can add AI players' });
-      return;
-    }
-
-    // Cannot add AI after game started
-    if (room.gameState?.phase === 'playing') {
-      this.transport.send(client.playerId, { type: 'error', message: 'Cannot add AI during game' });
-      return;
-    }
-
-    // Max 6 total players
-    const totalPlayers = room.clients.size + room.aiPlayers.size;
-    if (totalPlayers >= 6) {
-      this.transport.send(client.playerId, { type: 'error', message: 'Game is full' });
-      return;
-    }
-
-    const aiId = `ai${++this.aiCounter}`;
-    const usedNames = new Set([
-      ...Array.from(room.clients).map((id) => this.clients.get(id)?.playerName),
-      ...Array.from(room.aiPlayers.values()).map((ai) => ai.name),
-    ]);
-    const availableNames = AI_NAMES.filter((name) => !usedNames.has(name));
-    const aiName = availableNames.length > 0 ? availableNames[0] : `AI Bot ${this.aiCounter}`;
-
-    const aiPlayer: AiPlayer = {
-      id: aiId,
-      name: aiName,
-      gameId: room.id,
-      lastActionTime: 0,
-    };
-
-    room.aiPlayers.set(aiId, aiPlayer);
-    this.broadcastLobbyUpdate(room);
-    this.broadcastGameList();
-
-    console.log(`AI player ${aiName} added to game ${room.id}`);
-  }
-
-  private startGameLoop(room: GameRoom): void {
-    room.gameLoop = setInterval(() => {
-      if (!room.gameState || room.gameState.phase !== 'playing') return;
-
-      const now = Date.now();
-
-      // Process AI actions
-      this.processAiActions(room, now);
-
-      room.gameState.countries = room.gameState.countries.map((country: Country) => {
-        if (now >= country.nextBattleTime) {
-          return processBattle(country, room.gameState!.countries, now);
-        }
-        return country;
-      });
-
-      room.gameState.companies = room.gameState.companies.map((company: Company) => {
-        if (now >= company.nextUpdateTime) {
-          return updateStockPrice(company, now);
-        }
-        return company;
-      });
-
-      room.gameState.unitCost = calculateUnitCost(room.gameState.countries);
-
-      const winner = checkWinner(room.gameState.countries, room.gameState.factions);
-      if (winner) {
-        room.gameState.winner = winner;
-        room.gameState.phase = 'ended';
-        if (room.gameLoop) {
-          clearInterval(room.gameLoop);
-          room.gameLoop = null;
-        }
-        this.broadcastGameList();
-      }
-
-      this.transport.broadcastToGame(room, { type: 'gameState', state: room.gameState });
-    }, GAME.SERVER_TICK_RATE);
-  }
-
-  private processAiActions(room: GameRoom, now: number): void {
-    if (!room.gameState) return;
-
-    for (const aiPlayer of room.aiPlayers.values()) {
-      // Scale AI action interval based on unit cost to make it fairer
-      // As unit cost increases, AI acts slower (simulating having to "earn" money)
-      const scaledInterval = AI_ACTION_INTERVAL * (room.gameState.unitCost / GAME.BASE_UNIT_COST);
-
-      // Check if it's time for AI to take action
-      if (now - aiPlayer.lastActionTime < scaledInterval) continue;
-
-      // Find the AI's faction in the game
-      const player: Player | undefined = room.gameState.players.find(
-        (p: Player) => p.id === aiPlayer.id
-      );
-      if (!player) continue;
-
-      const factionId = player.factionId;
-
-      // Find valid placement locations for AI
-      const validLocations = room.gameState.countries.filter((country) =>
-        canPlaceUnits(country, room.gameState!.countries, factionId)
-      );
-
-      if (validLocations.length === 0) continue;
-
-      // Strategy: Prioritize contested territories, then expansion, then reinforcement
-      const contestedLocations = validLocations.filter((country) => {
-        const factionIds = Object.keys(country.units).filter(
-          (id) => country.units[id] > 0 && id !== 'neutral'
-        );
-        return factionIds.length > 1;
-      });
-
-      const expansionLocations = validLocations.filter((country) => {
-        const myUnits = country.units[factionId] || 0;
-        return myUnits === 0;
-      });
-
-      const reinforcementLocations = validLocations.filter((country) => {
-        const myUnits = country.units[factionId] || 0;
-        return myUnits > 0;
-      });
-
-      let targetCountry: Country;
-
-      if (contestedLocations.length > 0) {
-        // Reinforce contested territories
-        targetCountry = contestedLocations[Math.floor(Math.random() * contestedLocations.length)];
-      } else if (expansionLocations.length > 0 && Math.random() < 0.7) {
-        // Expand to new territories 70% of the time when available
-        targetCountry = expansionLocations[Math.floor(Math.random() * expansionLocations.length)];
-      } else if (reinforcementLocations.length > 0) {
-        // Reinforce existing territories
-        targetCountry =
-          reinforcementLocations[Math.floor(Math.random() * reinforcementLocations.length)];
-      } else {
-        // Fallback to any valid location
-        targetCountry = validLocations[Math.floor(Math.random() * validLocations.length)];
-      }
-
-      // Place units
-      const countryIndex = room.gameState.countries.findIndex(
-        (c) => c.coords.q === targetCountry.coords.q && c.coords.r === targetCountry.coords.r
-      );
-
-      if (countryIndex !== -1) {
-        const country = room.gameState.countries[countryIndex];
-        const newUnits = { ...country.units };
-        newUnits[factionId] = (newUnits[factionId] || 0) + 1;
-
-        room.gameState.countries[countryIndex] = {
-          ...country,
-          units: newUnits,
-          nextBattleTime: Date.now() + GAME.PLACEMENT_GRACE_PERIOD,
-        };
-
-        room.gameState.unitCost = calculateUnitCost(room.gameState.countries);
-        aiPlayer.lastActionTime = now;
-      }
-    }
-  }
-
-  getGamePlayers(room: GameRoom): Player[] {
-    const players: Player[] = [];
-    let factionIndex = 0;
-
-    for (const playerId of room.clients) {
-      const client = this.clients.get(playerId);
-      if (client?.playerName) {
-        players.push({
-          id: client.playerId,
-          name: client.playerName,
-          factionId: `faction${factionIndex++}`,
-          isAi: false,
-        });
-      }
-    }
-
-    // Add AI players
-    for (const aiPlayer of room.aiPlayers.values()) {
-      players.push({
-        id: aiPlayer.id,
-        name: aiPlayer.name,
-        factionId: `faction${factionIndex++}`,
-        isAi: true,
-      });
-    }
-
-    return players;
-  }
-
-  private broadcastLobbyUpdate(room: GameRoom): void {
-    const players = this.getGamePlayers(room);
-    this.transport.broadcastToGame(room, { type: 'lobbyUpdate', players });
-  }
-
-  broadcastGameList(): void {
-    if (this.onGameListChanged) {
-      this.onGameListChanged();
-    }
-  }
-
-  getGameList(): GameInfo[] {
-    const games: GameInfo[] = [];
-    for (const [id, room] of this.games) {
-      const host = this.clients.get(room.hostId);
-      const players = this.getGamePlayers(room);
-      games.push({
-        id,
-        hostName: host?.playerName ?? 'Unknown',
-        playerCount: players.length,
-        maxPlayers: 6,
-        phase: room.gameState?.phase ?? 'lobby',
-        players: players.map((p) => p.name),
-      });
-    }
-    return games;
-  }
-
-  destroyAllGames(): void {
-    for (const room of this.games.values()) {
-      if (room.gameLoop) {
-        clearInterval(room.gameLoop);
-      }
-    }
-    this.games.clear();
-  }
-}
